requirements.txt
stravalib
pandas
polars
plotly
ipykernel
xgboost
ipython
requests
python-dotenv
ruamel-yaml
duckdb
tqdm
pyarrow
